separated reviews and review api
