+++ app.js
@@ -8,6 +8,7 @@ var bodyParser = require('body-parser');
 var routes = require('./routes/index');
 var users = require('./routes/users');
 var reviews = require('./routes/reviews');
+var review = require('./routes/review');
 
 var database = require('./database/index');
 
@@ -28,6 +29,7 @@ app.use(express.static(path.join(__dirname, 'public')));
 app.use('/', routes);
 app.use('/users', users);
 app.use('/reviews', reviews);
+app.use('/review', review);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
