+++ database/index.js
@@ -1,6 +1,8 @@
 'use strict'
 var mongoose = require('mongoose');
-var mongolabStringConnexion = 'mongodb://localhost:27017/ws-restapi';
+
+var mongolabStringConnexion = process.env.MONGO_STRING_CONNECTION || 'mongodb://localhost:27017/ws-restapi';
+
 mongoose.connect(mongolabStringConnexion);
 var db = mongoose.connection;
 db.on('error', console.error.bind(console, 'connection error:'));
