+++ routes/reviews.js
@@ -4,7 +4,7 @@ var router = express.Router();
 var reviewsDb = require('../database/reviews');
 
 /* API GET reviews page. */
-router.route('/api').get(function (req, res) {
+router.route('/').get(function (req, res) {
 	reviewsDb.find({}, function (err, reviews) {
 		if (err) {
 			res.status(500).send({'error': err});
@@ -14,8 +14,19 @@ router.route('/api').get(function (req, res) {
 	});
 });
 
+/* API DELETE all reviews. */
+router.route('/').delete(function (req, res) {
+	reviewsDb.remove( function (err, review) {
+		if (err) {
+			res.status(500).send({'error': err});
+		} else {
+			res.status(204);
+		}
+	});
+});
+
 /* API POST review */
-router.route('/api').post(function (req, res) {
+router.route('/').post(function (req, res) {
 	if(req.body.name === undefined || req.body.placeType === undefined || req.body.stars === undefined) {
 		res.status(400).send("bad attributes");
 	} else {
@@ -33,51 +44,6 @@ router.route('/api').post(function (req, res) {
 			}
 		});
 	}
-});
-
-/* API GET one review. */
-router.route('/api/:id').get(function (req, res) {
-	var id = req.params.id;
-	reviewsDb.findById(id, function (err, review) {
-		if (err) {
-			res.status(404).send({'error': err});
-		} else {
-			res.send(review);
-		}
-	});
-});
-
-/* API PUT review. */
-router.route('/api/:id').put(function (req, res) {
-	if(req.body.name === undefined || req.body.placeType === undefined || req.body.stars === undefined) {
-		res.status(400).send("bad attributes");
-	} else {
-	  var newReview= {
-	  					name: req.body.name,
-	  					placeType: req.body.placeType,
-	  					stars: req.body.stars
-	  				 }
-	  var id= req.params.id;
-	  reviewsDb.findByIdAndUpdate(id, newReview, function (err, review) {
-		if (err) {
-			res.status(404).send({'error': err});
-		} else {
-			res.send(newReview);
-		}
-	});
-	}
-});
-
-/* API GET one review. */
-router.route('/api/:id').delete(function (req, res) {
-	var id = req.params.id;
-	reviewsDb.findByIdAndRemove(id, function (err, review) {
-		if (err) {
-			res.status(500).send({'error': err});
-		} else {
-			res.send("review deleted");
-		}
-	});
 });
 
 module.exports = router;
